+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-09-10
+### Feat
+- Add Variables class for shareable config values such as base styles for success / warning / info as well as icon paths
+- Add Action button with quick access to static predefined buttons (+ / - / edit / search...)
+- Add Row component
+- Add Typography component with static methods for Headings
+- Add Editor Window to present the different components `Slax > UI Toolkit Demo`
+
 ## [0.1.0] - 2024-09-08
 ### Feat
 - Add AlertBox component with factory methods to create success / warning / info / danger boxes quickly.